Refactor and use our reviewdog fork
We have reusable components already set up that point where we want.
Removing a bunch of the stuff here lets those do their thing without us
needing to copy around a bunch of code.

The main thing that's purely trimmed out is shellcheck.  It's not
necessary for actionlint to run, and in practice it will already
actually be installed generally as part of the set of linters.  But I'm
not sure if we should perhaps have it here too.